refactor: move console attachment in a separate file
Separate the logic to attach stdin and stdout in utils.go in order to be
able to reuse it in other functions in virtctl.
