+++ pkg/virtctl/utils/BUILD.bazel
@@ -1,0 +1,9 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["utils.go"],
+    importpath = "kubevirt.io/kubevirt/pkg/virtctl/utils",
+    visibility = ["//visibility:public"],
+    deps = ["//vendor/golang.org/x/crypto/ssh/terminal:go_default_library"],
+)
